+++ cmd/root.go
@@ -12,6 +12,11 @@ var (
 func init() {
 	cli.SetShort("Lu(cy) CLI")
 	cli.SetLong("Named after myself, because I'm vain.\nPronounced 'loosely', because while it's cool, its utility is only loosely based in reality.")
+
+	// TODO: global flag: creds
+
+	cli.Flags().StringP("aws-profile", "p", "lmhd_root", "AWS Profile to use (if applicable)")
+	cli.BindFlags()
 }
 
 func Execute() {